test: use symlinked app directory directly
We used to copy the app content, to avoid the tests accidentally
mutating it.

But it makes it difficult to change the source code and having it
reflected inside the VM. Also Yunohost copies the app files itself
before executing scripts. So this should be more convenient and safe
enough.